requirements.txt
Flask==0.10.1
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
protobuf==3.0.0b2
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyYAML==3.11
redis==2.10.5
rsa==3.3
six==1.10.0
Werkzeug==0.11.3
wheel==0.24.0
